requirements.txt
# Core ML and Data Handling
pandas
scikit-learn
numpy
kagglehub
pyarrow
fastparquet

# Real-time Data Pipeline
kafka-python
scapy

# Backend API Framework
fastapi[all]
uvicorn

# Database
SQLAlchemy[asyncio]
psycopg2-binary

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-jwt

# Notifications & External Services
twilio
python-dotenv
google-generativeai
requests